ENH: add logging system that works across multiple processes
Logging system allows subprocesses to send logs to a central listener
in the main process, writes logs to both logfile and the terminal.

Logs files are written to the log dir specified in the user's config file or the settings
dialog menu. Logs are appending to daily logfiles.

We make use of the logging.handlers classes from the python standard library, mainly:
    QueueListener (in main process) — collects log records from a multiprocessing.Queue
    QueueHandler (in subprocesses) — sends log records to the main process queue

For example usage (in a simple context), see src/badger/tests/multiprocess_logging.py